更新商品库应用模块

+++ content/apps/goodslib/admin.php
@@ -85,6 +85,7 @@ class admin extends ecjia_admin {
         RC_Loader::load_app_class('goods', 'goods', false );
         RC_Loader::load_app_class('goodslib', 'goodslib' );
         RC_Loader::load_app_class('goods_image_data', 'goodslib', false);
+        RC_Loader::load_app_class('product_image_data', 'goodslib', false);
 //         RC_Loader::load_app_class('goods_imageutils', 'goodslib', false);
         
         RC_Loader::load_app_func('admin_category', 'goods');
@@ -109,6 +110,7 @@ class admin extends ecjia_admin {
         $this->db_goods = RC_Model::model('goods/goods_model');
         
         ecjia_admin_log::instance()->add_object('goodslib', __('商品库', 'goodslib'));
+        ecjia_admin_log::instance()->add_object('goodslib_product', __('商品库货品', 'goodslib'));
     }
     
     /**
@@ -119,7 +121,7 @@ class admin extends ecjia_admin {
         
         $cat_id = empty($_GET['cat_id']) ? 0 : intval($_GET['cat_id']);
         
-        $this->assign('ur_here', __('商品库', 'goodslib'));
+        $this->assign('ur_here', __('商品库商品（SPU）', 'goodslib'));
         ecjia_screen::get_current_screen()->add_nav_here(new admin_nav_here(__('商品库', 'goodslib')));
         
         $this->assign('cat_list', cat_list(0, $cat_id, false));
@@ -924,7 +926,7 @@ class admin extends ecjia_admin {
         $goods = RC_DB::table('goodslib')->where('goods_id', $goods_id)->where('is_delete', 0)->first();
         
         if (empty($goods)) {
-            return $this->showmessage(__('未检测到此商品', 'goodslib'), ecjia::MSGTYPE_HTML | ecjia::MSGSTAT_ERROR, array('links' => array(array('text'=> __('返回商品列表', 'goodslib'),'href'=>RC_Uri::url('goods/merchant/init')))));
+            return $this->showmessage(__('未检测到此商品', 'goodslib'), ecjia::MSGTYPE_HTML | ecjia::MSGSTAT_ERROR, array('links' => array(array('text'=> __('返回商品列表', 'goodslib'),'href'=>RC_Uri::url('goodslib/admin/init')))));
         }
         
         if (!empty($goods['goods_desc'])) {
@@ -1160,8 +1162,8 @@ class admin extends ecjia_admin {
      * 商品描述更新
      */
     public function update_goods_desc() {
-        
-        $goods_type = isset($_POST['goods_type']) ? $_POST['goods_type'] : 0;
+        $this->admin_priv('goodslib_update');
+
         $goods_id = intval($_REQUEST['goods_id']);
         
         $goods = RC_DB::table('goodslib')->where('goods_id', $goods_id)->first();
@@ -1312,6 +1314,11 @@ class admin extends ecjia_admin {
                             RC_DB::table('goodslib_attr')->where('goods_attr_id', $info['goods_attr_id'])->update($data);
                         }
                     } else {
+                        //查询要删除的属性是否存在对应货品
+                        $count = RC_DB::table('goodslib_products')->where('goods_id', $goods_id)->whereRaw(RC_DB::raw(" CONCAT('|', goods_attr, '|') like '%|".$info['goods_attr_id']."|%' "))->count();
+                        if($count) {
+                            return $this->showmessage(__('请先删除关联的货品再修改规格属性', 'goodslib'), ecjia::MSGTYPE_JSON | ecjia::MSGSTAT_ERROR);
+                        }
                         RC_DB::table('goodslib_attr')->where('goods_attr_id', $info['goods_attr_id'])->delete();
                     }
                 }
@@ -1379,12 +1386,273 @@ class admin extends ecjia_admin {
         $this->assign('form_action', 		RC_Uri::url('goodslib/admin/product_add_execute'));
         if (isset($_GET['step'])) {
             $this->assign('step', 8);
-            $ur_here = __('添加货品', 'goodslib');
+            $ur_here = __('添加货品（SKU）', 'goodslib');
         } else {
-            $ur_here = __('编辑货品', 'goodslib');
+            $ur_here = __('编辑货品（SKU）', 'goodslib');
         }
         $this->assign('ur_here', $ur_here);
         $this->display('product_info.dwt');
+    }
+
+    public function product_edit() {
+        $this->admin_priv('goodslib_manage');
+
+        $goods_id = intval($_GET['goods_id']);
+        $product_id = intval($_GET['id']);
+        ecjia_screen::get_current_screen()->add_nav_here(new admin_nav_here(__('商品库', 'goodslib'), RC_Uri::url('goodslib/admin/init')));
+        ecjia_screen::get_current_screen()->add_nav_here(new admin_nav_here(__('货品列表', 'goodslib'), RC_Uri::url('goodslib/admin/product_list', 'goods_id='.$goods_id)));
+        $ur_here = __('编辑货品（SKU）', 'goodslib');
+        $this->assign('ur_here', $ur_here);
+
+        ecjia_screen::get_current_screen()->add_nav_here(new admin_nav_here($ur_here));
+
+
+        $db_goods = RC_DB::table('goodslib')->where('goods_id', $goods_id);
+
+        //获取商品的信息
+        $goods = $db_goods->select('goods_sn', 'goods_name', 'goods_type', 'shop_price')->first();
+        if (empty($goods)) {
+            return $this->showmessage(__('未检测到此商品', 'goodslib'), ecjia::MSGTYPE_HTML | ecjia::MSGSTAT_ERROR, array('links' => array(array('text' => __('返回商品列表', 'goodslib'), 'href' => RC_Uri::url('goodslib/admin/init')))));
+        }
+
+        $info = RC_DB::table('goodslib_products')->where('product_id', $product_id)->where('goods_id', $goods_id)->first();
+        /* 商品图片路径 */
+        if (!empty($info['product_img'])) {
+            $info['product_img'] 	= goods_imageutils::getAbsoluteUrl($info['product_img']);
+            $info['product_thumb'] 	= goods_imageutils::getAbsoluteUrl($info['product_thumb']);
+            $info['product_original_img'] 	= goods_imageutils::getAbsoluteUrl($info['product_original_img']);
+        }
+
+        $this->assign('action_link', array('href' => RC_Uri::url('goodslib/admin/product_list', ['goods_id' => $goods_id]), 'text' => __('商品编辑', 'goodslib')));
+
+
+        $this->assign('goods_name', 		sprintf(__('商品名称：%s', 'goodslib'), $goods['goods_name']));
+        $this->assign('goods', $goods);
+        $this->assign('info', $info);
+        $this->assign('form_action', 		RC_Uri::url('goodslib/admin/product_update'));
+
+        $product = goodslib_product_list($goods_id, '');
+        $this->assign('product_list', $product['product']);
+
+        $this->display('product_edit.dwt');
+    }
+
+    public function product_update()
+    {
+        $this->admin_priv('goodslib_manage');
+
+        $product_id = !empty($_POST['product_id']) ? intval($_POST['product_id']) : 0;
+        $info = RC_DB::table('goodslib_products')->where('product_id', $product_id)->first();
+        $goods = RC_DB::table('goodslib')->where('goods_id', $info['goods_id'])->first();
+
+        /* 检查货号是否重复 */
+        if (trim($_POST['product_sn'])) {
+            $count = RC_DB::table('goodslib_products')->where('product_sn', trim($_POST['product_sn']))->where('product_id', '!=', $product_id)->count();
+            if ($count > 0) {
+                return $this->showmessage(__('您输入的货号已存在，请换一个', 'goodslib'), ecjia::MSGTYPE_JSON | ecjia::MSGSTAT_ERROR);
+            }
+        }
+
+        $upload = RC_Upload::uploader('image', array('save_path' => 'images', 'auto_sub_dirs' => true));
+        $upload->add_saving_callback(function ($file, $filename) {
+            return true;
+        });
+
+        /* 是否处理商品图 */
+        $proc_goods_img = true;
+        if (isset($_FILES['goods_img']) && !$upload->check_upload_file($_FILES['goods_img'])) {
+            $proc_goods_img = false;
+        }
+//
+        /* 是否处理缩略图 */
+        $proc_thumb_img = isset($_FILES['thumb_img']) ? true : false;
+        if (isset($_FILES['thumb_img']) && !$upload->check_upload_file($_FILES['thumb_img'])) {
+            $proc_thumb_img = false;
+        }
+
+        if ($proc_goods_img) {
+            if (isset($_FILES['goods_img'])) {
+                $image_info = $upload->upload($_FILES['goods_img']);
+                if (empty($image_info)) {
+                    return $this->showmessage($upload->error(), ecjia::MSGTYPE_JSON | ecjia::MSGSTAT_ERROR);
+                }
+            }
+        }
+
+        if ($proc_thumb_img) {
+            if (isset($_FILES['thumb_img'])) {
+                $thumb_info = $upload->upload($_FILES['thumb_img']);
+                if (empty($thumb_info)) {
+                    return $this->showmessage($upload->error(), ecjia::MSGTYPE_JSON | ecjia::MSGSTAT_ERROR);
+                }
+            }
+        }
+
+        /* 处理商品图片 */
+        $goods_img	 	= ''; // 初始化商品图片
+        $goods_thumb 	= ''; // 初始化商品缩略图
+        $img_original	= ''; // 初始化原始图片
+
+        /* 处理商品数据 */
+        $product_name 	= !empty($_POST['product_name']) 		? $_POST['product_name'] 				: '';
+        $shop_price 	= $_POST['product_shop_price'] != '' 	? $_POST['product_shop_price'] 				: NULL;
+        $product_sn     = !empty($_POST['product_sn'])          ? trim($_POST['product_sn'])                : '';
+
+        $product_bar_code = isset($_POST['product_bar_code']) 	? $_POST['product_bar_code'] 	: '';
+
+        //货品号不为空
+        if (!empty($product_sn)) {
+            /* 检测：货品货号是否在商品表和货品表中重复 */
+            if (check_goodslib_sn_exist($product_sn, $info['goods_id'])) {
+                return $this->showmessage(__('货品号已存在，请修改', 'goodslib'), ecjia::MSGTYPE_JSON | ecjia::MSGSTAT_ERROR);
+            }
+            if (check_goodslib_product_sn_exist($product_sn, $product_id)) {
+                return $this->showmessage(__('货品号已存在，请修改', 'goodslib'), ecjia::MSGTYPE_JSON | ecjia::MSGSTAT_ERROR);
+            }
+        } else {
+            //货品号为空 自动补货品号
+            $product_sn = $goods['goods_sn'] . "g_p" . $product_id;
+        }
+
+        $data = array(
+            'product_name'				=> rc_stripslashes($product_name),
+            'product_sn'			  	=> $product_sn,
+            'product_shop_price'		=> $shop_price,
+            'product_bar_code'		  	=> $product_bar_code,
+        );
+        RC_DB::table('goodslib_products')->where('product_id', $product_id)->update($data);
+
+        /* 记录日志 */
+        $log_object = $product_name;
+        if(empty($log_object)) {
+            $log_object = $goods['goods_name'];
+        }
+        ecjia_admin::admin_log($log_object, 'edit', 'goodslib_product');
+
+        /* 更新上传后的商品图片 */
+        if ($proc_goods_img) {
+            if (isset($image_info)) {
+                $goods_image = new product_image_data($image_info['name'], $image_info['tmpname'], $image_info['ext'], $info['goods_id'], $product_id);
+                if ($proc_thumb_img) {
+                    $goods_image->set_auto_thumb(false);
+                }
+
+                $result = $goods_image->update_goods();
+
+                if (is_ecjia_error($result)) {
+                    return $this->showmessage($result->get_error_message(), ecjia::MSGTYPE_JSON | ecjia::MSGSTAT_ERROR);
+                }
+            }
+        }
+
+        /* 更新上传后的缩略图片 */
+        if ($proc_thumb_img) {
+            if (isset($thumb_info)) {
+                $thumb_image = new product_image_data($thumb_info['name'], $thumb_info['tmpname'], $thumb_info['ext'], $info['goods_id'], $product_id);
+                $result = $thumb_image->update_thumb();
+                if (is_ecjia_error($result)) {
+                    return $this->showmessage($result->get_error_message(), ecjia::MSGTYPE_JSON | ecjia::MSGSTAT_ERROR);
+                }
+            }
+        }
+
+        $pjaxurl = RC_Uri::url('goodslib/admin/product_edit', "id=$product_id&goods_id=".$info['goods_id']);
+        return $this->showmessage(__('编辑货品成功', 'goodslib'), ecjia::MSGTYPE_JSON | ecjia::MSGSTAT_SUCCESS, array('pjaxurl' => $pjaxurl));
+
+    }
+
+    /**
+     * 货品编辑
+     */
+    public function product_desc_edit()
+    {
+        $this->admin_priv('goodslib_manage');
+
+        $goods_id = intval($_GET['goods_id']);
+        $product_id = intval($_GET['id']);
+        ecjia_screen::get_current_screen()->add_nav_here(new admin_nav_here(__('商品库', 'goodslib'), RC_Uri::url('goodslib/admin/init')));
+        ecjia_screen::get_current_screen()->add_nav_here(new admin_nav_here(__('货品列表', 'goodslib'), RC_Uri::url('goodslib/admin/product_list', 'goods_id='.$goods_id)));
+        $ur_here = __('编辑货品（SKU）', 'goodslib');
+        $this->assign('ur_here', $ur_here);
+
+        ecjia_screen::get_current_screen()->add_nav_here(new admin_nav_here($ur_here));
+
+
+        $db_goods = RC_DB::table('goodslib')->where('goods_id', $goods_id);
+
+        //获取商品的信息
+        $goods = $db_goods->select('goods_sn', 'goods_name', 'goods_type', 'shop_price')->first();
+        if (empty($goods)) {
+            return $this->showmessage(__('未检测到此商品', 'goodslib'), ecjia::MSGTYPE_HTML | ecjia::MSGSTAT_ERROR, array('links' => array(array('text' => __('返回商品列表', 'goodslib'), 'href' => RC_Uri::url('goods/merchant/init')))));
+        }
+
+        $info = RC_DB::table('goodslib_products')->where('product_id', $product_id)->where('goods_id', $goods_id)->first();
+        if (empty($info)) {
+            return $this->showmessage(__('未检测到此货品', 'goodslib'), ecjia::MSGTYPE_HTML | ecjia::MSGSTAT_ERROR, array('links' => array(array('text' => __('返回商品列表', 'goodslib'), 'href' => RC_Uri::url('goods/merchant/init')))));
+        }
+
+        $this->assign('action_link', array('href' => RC_Uri::url('goodslib/admin/product_list', ['goods_id' => $goods_id]), 'text' => __('商品编辑', 'goodslib')));
+
+
+        $this->assign('goods_name', 		sprintf(__('商品名称：%s', 'goodslib'), $goods['goods_name']));
+        $this->assign('goods', $goods);
+        $this->assign('nav_tag', 'desc');
+
+
+        if (!empty($info['product_desc'])) {
+            $info['product_desc'] = stripslashes($info['product_desc']);
+        }
+        $this->assign('info', $info);
+        $this->assign('goods_id', $goods_id);
+        $this->assign('form_action', RC_Uri::url('goodslib/admin/product_desc_update', 'goods_id='.$goods_id.'&id='.$product_id));
+
+        $product = goodslib_product_list($goods_id, '');
+        $this->assign('product_list', $product['product']);
+
+        $this->display('product_desc.dwt');
+    }
+    /**
+     * 货品描述更新
+     */
+    public function product_desc_update() {
+        $this->admin_priv('goodslib_update', ecjia::MSGTYPE_JSON); // 检查权限
+
+
+        $goods_id 	= !empty($_GET['goods_id'])		? intval($_GET['goods_id']) : 0;
+        $product_id 	= !empty($_GET['id'])		? intval($_GET['id']) : 0;
+
+        $db_goods = RC_DB::table('goodslib')->where('goods_id', $goods_id);
+
+        //获取商品的信息
+        $goods = $db_goods->select('goods_sn', 'goods_name', 'goods_type', 'shop_price')->first();
+        if (empty($goods)) {
+            return $this->showmessage(__('未检测到此商品', 'goodslib'), ecjia::MSGTYPE_HTML | ecjia::MSGSTAT_ERROR, array('links' => array(array('text' => __('返回商品列表', 'goodslib'), 'href' => RC_Uri::url('goodslib/admin/init')))));
+        }
+
+        $info = RC_DB::table('goodslib_products')->where('product_id', $product_id)->where('goods_id', $goods_id)->first();
+        if (empty($info)) {
+            return $this->showmessage(__('未检测到此货品', 'goodslib'), ecjia::MSGTYPE_HTML | ecjia::MSGSTAT_ERROR, array('links' => array(array('text' => __('返回商品列表', 'goodslib'), 'href' => RC_Uri::url('goodslib/admin/init')))));
+        }
+
+        $data = array(
+            'product_desc'  => !empty($_POST['product_desc']) ? $_POST['product_desc'] : '',
+        );
+        RC_DB::table('goodslib_products')->where('product_id', $product_id)->where('goods_id', $goods_id)->update($data);
+
+        /* 记录日志 */
+        $log_object = $info['product_name'];
+        if(empty($log_object)) {
+            $log_object = $goods['goods_name'];
+        }
+        ecjia_admin::admin_log($log_object, 'edit', 'goodslib_product');
+
+        /* 提示页面 */
+
+        $arr['goods_id'] = $goods_id;
+        $arr['id'] = $product_id;
+        $message = __('编辑成功', 'goodslib');
+        $url = RC_Uri::url('goodslib/admin/product_desc_edit', $arr);
+        return $this->showmessage($message, ecjia::MSGTYPE_JSON | ecjia::MSGSTAT_SUCCESS, array('url' => $url, 'goods_id' => $goods_id,'id' => $product_id));
     }
     
     /**
@@ -1554,9 +1822,14 @@ class admin extends ecjia_admin {
         $this->admin_priv('goodslib_update', ecjia::MSGTYPE_JSON);
         
         $product_id = !empty($_GET['id']) ? intval($_GET['id']) : 0;
-        
+
+        $image_data = new product_image_data('', '', '', 0, $product_id);
+        $image_data->delete_images();
+
         $result = RC_DB::table('goodslib_products')->where('product_id', $product_id)->delete();
         if ($result) {
+            $image_data->delete_gallery();
+            RC_DB::table('goodslib_gallery')->where('product_id', $product_id)->delete();
             //记录日志
             ecjia_admin::admin_log('', 'trash', 'products');
         }
